+++ frontend/src/components/FormContainer.tsx
@@ -8,6 +8,7 @@ import SubmittedPage from "./SubmittedPage";
 import { Route, Switch, useLocation, useSearchParams } from "wouter";
 import { getConfig } from "../config/configLoader";
 import EditPage from "./EditPage";
+import { fnv1a32 } from "../fnv1a32";
 
 export interface FormData extends Record<string, string> {
   mainProblem: string;
@@ -78,16 +79,6 @@ const usePreviousLocation = () => {
 
   return previousLocation;
 };
-
-// hash string to 32 bit integer
-export function fnv1a32(str: string): string {
-  let h = 0x811c9dc5 >>> 0;
-  for (let i = 0; i < str.length; i++) {
-    h ^= str.charCodeAt(i);
-    h = (h + ((h << 1) + (h << 4) + (h << 7) + (h << 8) + (h << 24))) >>> 0;
-  }
-  return (h >>> 0).toString(16);
-}
 
 const FormContainer = () => {
   const config = getConfig();
