+++ frontend/src/fnv1a32.ts
@@ -1,0 +1,9 @@
+// hash string to 32 bit integer
+export function fnv1a32(str: string): string {
+  let h = 0x811c9dc5 >>> 0;
+  for (let i = 0; i < str.length; i++) {
+    h ^= str.charCodeAt(i);
+    h = (h + ((h << 1) + (h << 4) + (h << 7) + (h << 8) + (h << 24))) >>> 0;
+  }
+  return (h >>> 0).toString(16);
+}
